Revert "Update configs (#463)"
This reverts commit f708a696bb206d356c57a08b66509d60e36f29ec.
