+++ apps/cms/tsconfig.json
@@ -1,11 +1,9 @@
 {
   "$schema": "https://json.schemastore.org/tsconfig",
-  "extends": "tsconfig/base.json",
-
+  "extends": "../../tsconfig.json",
   "compilerOptions": {
     "allowJs": true
   },
-
-  "include": [".eslintrc.cjs", "*.ts", "src", "schemas"],
-  "exclude": ["node_modules", "dist", ".sanity", "migrations"]
+  "include": ["**/*.ts", "**/*.tsx"],
+  "exclude": ["node_modules", "dist", ".sanity"]
 }
