+++ apps/docs/package.json
@@ -11,22 +11,22 @@
     "type-check": "tsc --noEmit"
   },
   "dependencies": {
+    "@echo-webkom/tailwind-config": "0.1.0",
     "next": "13.4.12",
-    "nextra": "2.10.0",
-    "nextra-theme-docs": "2.10.0",
-    "react": "18.2.0",
-    "react-dom": "18.2.0"
+    "nextra": "latest",
+    "nextra-theme-docs": "latest",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
     "@types/node": "18.17.0",
     "@types/react": "18.2.15",
-    "tsconfig": "workspace:*",
     "@types/react-dom": "18.2.7",
     "autoprefixer": "10.4.14",
-    "eslint": "8.45.0",
-    "eslint-config-webkom": "workspace:*",
-    "postcss": "8.4.27",
-    "tailwindcss": "3.3.3",
-    "typescript": "5.1.6"
+    "eslint": "^8.45.0",
+    "eslint-config-custom": "0.0.0",
+    "postcss": "^8.4.27",
+    "tailwindcss": "^3.3.3",
+    "typescript": "^5.1.6"
   }
 }
