+++ apps/docs/postcss.config.cjs
@@ -1,8 +1,2 @@
-const config = {
-  plugins: {
-    tailwindcss: {},
-    autoprefixer: {},
-  },
-};
-
-module.exports = config;
+// @ts-ignore
+module.exports = require("@echo-webkom/tailwind-config/postcss");
