+++ apps/docs/tailwind.config.ts
@@ -3,8 +3,5 @@ import type {Config} from "tailwindcss";
 export default {
   darkMode: "class",
   content: ["./src/**/*.{tsx,mdx}", "./theme.config.tsx"],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
+  presets: [require("@echo-webkom/tailwind-config")],
 } satisfies Config;
