+++ apps/docs/tsconfig.json
@@ -1,7 +1,6 @@
 {
   "$schema": "https://json.schemastore.org/tsconfig",
-  "extends": "tsconfig/base.json",
-
+  "extends": "../../tsconfig.json",
   "compilerOptions": {
     "strictNullChecks": true,
     "baseUrl": ".",
@@ -18,15 +17,13 @@
   "mdx": {
     "plugins": ["remark-gfm"]
   },
-
   "include": [
     "next-env.d.ts",
-    ".eslintrc.cjs",
     "**/*.ts",
     "**/*.tsx",
     "**/*.cjs",
     "**/*.mjs",
     ".next/types/**/*.ts"
   ],
-  "exclude": ["node_modules"]
+  "exclude": []
 }
