+++ apps/web/package.json
@@ -16,6 +16,7 @@
     "@echo-webkom/auth": "0.1.0",
     "@echo-webkom/db": "0.1.0",
     "@echo-webkom/lib": "0.1.0",
+    "@echo-webkom/tailwind-config": "0.1.0",
     "@hookform/resolvers": "^3.1.1",
     "@next/bundle-analyzer": "^13.4.12",
     "@radix-ui/react-accordion": "1.1.2",
@@ -61,16 +62,12 @@
     "@types/node": "18.17.0",
     "@types/react": "18.2.15",
     "@types/react-dom": "18.2.7",
-    "@tailwindcss/forms": "0.5.4",
-    "@tailwindcss/typography": "0.5.9",
-    "tsconfig": "workspace:*",
     "autoprefixer": "10.4.14",
     "dotenv-cli": "7.2.1",
     "eslint": "^8.45.0",
-    "eslint-config-webkom": "workspace:*",
+    "eslint-config-custom": "0.0.0",
     "postcss": "^8.4.27",
     "tailwindcss": "^3.3.3",
-    "tailwindcss-animate": "1.0.6",
     "typescript": "^5.1.6"
   }
 }
