+++ package.json
@@ -36,13 +36,34 @@
   },
   "dependencies": {
     "@ianvs/prettier-plugin-sort-imports": "^4.1.0",
+    "@types/node": "18.17.0",
     "@types/prettier": "^2.7.3",
+    "eslint": "^8.45.0",
+    "eslint-config-custom": "0.0.0",
     "husky": "8.0.3",
+    "lint-staged": "13.2.3",
     "prettier": "2.8.8",
     "prettier-plugin-organize-imports": "3.2.3",
     "prettier-plugin-tailwindcss": "0.4.1",
     "start-server-and-test": "^2.0.0",
-    "turbo": "1.10.9"
+    "turbo": "1.10.9",
+    "typescript": "^5.1.6"
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "lint-staged"
+    }
+  },
+  "lint-staged": {
+    "*.{js,jsx,ts,tsx,md,mdx,mjs,cjs}": [
+      "eslint",
+      "prettier --check \"**/*.{ts,tsx,js,cjs,json,md,mdx}\""
+    ]
+  },
+  "pnpm": {
+    "overrides": {
+      "next@latest": "13.0.6"
+    }
   },
   "packageManager": "pnpm@8.6.9"
 }
