+++ packages/auth/package.json
@@ -7,7 +7,7 @@
   "scripts": {
     "clean": "rm -rf .turbo node_modules",
     "lint": "eslint .",
-    "lint:fix": "eslint lint --fix",
+    "lint:fix": "pnpm lint --fix",
     "type-check": "tsc --noEmit"
   },
   "dependencies": {
@@ -21,8 +21,7 @@
   "devDependencies": {
     "@types/node": "18.17.0",
     "eslint": "^8.45.0",
-    "eslint-config-webkom": "0.0.0",
-    "tsconfig": "workspace:*",
+    "eslint-config-custom": "0.0.0",
     "typescript": "^5.1.6"
   }
 }
