+++ packages/config/eslint/package.json
@@ -1,19 +1,19 @@
 {
-  "name": "eslint-config-webkom",
+  "name": "eslint-config-custom",
   "version": "0.0.0",
   "main": "index.js",
-  "private": true,
   "license": "MIT",
   "dependencies": {
     "@types/eslint": "^8.44.0",
-    "@types/node": "18.17.0",
     "@typescript-eslint/eslint-plugin": "^6.1.0",
-    "@typescript-eslint/parser": "^6.2.0",
-    "eslint": "^8.45.0",
+    "@typescript-eslint/parser": "^6.1.0",
     "eslint-config-next": "13.4.12",
     "eslint-config-prettier": "8.8.0",
     "eslint-config-turbo": "^1.10.9",
-    "eslint-plugin-react": "7.33.0",
+    "eslint-plugin-react": "7.33.0"
+  },
+  "devDependencies": {
+    "eslint": "^8.45.0",
     "typescript": "^5.1.6"
   }
 }
