+++ packages/config/tailwind/index.ts
@@ -1,0 +1,89 @@
+import {type Config} from "tailwindcss";
+import {fontFamily} from "tailwindcss/defaultTheme";
+
+module.exports = {
+  content: ["./src/**/*.{ts,js,tsx,jsx}"],
+  theme: {
+    fontFamily: {
+      primary: ["var(--inter-font)", ...fontFamily.sans],
+      sans: ["var(--inter-font)", ...fontFamily.sans],
+      mono: ["var(--ibm-font)", ...fontFamily.mono],
+      display: ["var(--inter-display-font)", ...fontFamily.sans],
+    },
+    extend: {
+      colors: {
+        background: "var(--background)",
+        foreground: "var(--foreground)",
+        border: "var(--border)",
+        input: "var(--input)",
+        ring: "var(--ring)",
+        wave: "var(--wave)",
+        banner: "var(--banner)",
+        primary: {
+          DEFAULT: "var(--primary)",
+          foreground: "var(--primary-foreground)",
+        },
+        secondary: {
+          DEFAULT: "var(--secondary)",
+          foreground: "var(--secondary-foreground)",
+        },
+        destructive: {
+          DEFAULT: "var(--destructive)",
+          foreground: "var(--destructive-foreground)",
+        },
+        muted: {
+          DEFAULT: "var(--muted)",
+          foreground: "var(--muted-foreground)",
+        },
+        accent: {
+          DEFAULT: "var(--accent)",
+          foreground: "var(--accent-foreground)",
+        },
+        success: {
+          DEFAULT: "var(--success)",
+          foreground: "var(--success-foreground)",
+        },
+        info: {
+          DEFAULT: "var(--info)",
+          foreground: "var(--info-foreground)",
+        },
+        warning: {
+          DEFAULT: "var(--warning)",
+          foreground: "var(--warning-foreground)",
+        },
+        popover: {
+          DEFAULT: "var(--popover)",
+          foreground: "var(--popover-foreground)",
+        },
+        card: {
+          DEFAULT: "var(--card)",
+          foreground: "var(--card-foreground)",
+        },
+      },
+      borderRadius: {
+        lg: "var(--radius)",
+        md: "calc(var(--radius) - 2px)",
+        sm: "calc(var(--radius) - 4px)",
+      },
+      keyframes: {
+        "accordion-down": {
+          from: {height: "0"},
+          to: {height: "var(--radix-accordion-content-height)"},
+        },
+        "accordion-up": {
+          from: {height: "var(--radix-accordion-content-height)"},
+          to: {height: "0"},
+        },
+      },
+      animation: {
+        "accordion-down": "accordion-down 0.2s ease-out",
+        "accordion-up": "accordion-up 0.2s ease-out",
+      },
+    },
+  },
+  plugins: [
+    require("tailwindcss-animate"),
+    require("@tailwindcss/forms"),
+    require("@tailwindcss/typography"),
+  ],
+} satisfies Config;
