+++ packages/config/tailwind/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "@echo-webkom/tailwind-config",
+  "version": "0.1.0",
+  "main": "index.js",
+  "license": "MIT",
+  "files": [
+    "index.js",
+    "postcss.js"
+  ],
+  "dependencies": {
+    "@tailwindcss/forms": "0.5.4",
+    "@tailwindcss/typography": "0.5.9",
+    "autoprefixer": "10.4.14",
+    "postcss": "^8.4.27",
+    "tailwindcss": "^3.3.3",
+    "tailwindcss-animate": "1.0.6",
+    "typescript": "^5.1.6"
+  }
+}
