+++ packages/config/tailwind/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "$schema": "https://json.schemastore.org/tsconfig",
+  "extends": "../../../tsconfig.json",
+  "exclude": ["index.ts"]
+}
