+++ packages/lib/tsconfig.json
@@ -1,5 +1,4 @@
 {
-  "extends": "tsconfig/base.json",
-  "include": [".eslintrc.cjs", "src", "index.ts"],
-  "exclude": ["node_modules"]
+  "extends": "../../tsconfig.json",
+  "include": ["src", "index.ts"]
 }
