--- packages/tsconfig/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "tsconfig",
-  "version": "0.0.0"
-}
