+++ tests/tsconfig.json
@@ -1,7 +1,6 @@
 {
   "$schema": "https://json.schemastore.org/tsconfig",
-  "extends": "tsconfig/base.json",
-
-  "include": [".eslintrc.cjs", "**/*.ts"],
+  "extends": "../tsconfig.json",
+  "include": ["**/*.ts"],
   "exclude": ["node_modules"]
 }
