+++ tsconfig.json
@@ -1,9 +1,9 @@
 {
   "$schema": "https://json.schemastore.org/tsconfig",
-  "display": "Base",
   "compilerOptions": {
-    "target": "ES2018",
+    "target": "es2017",
     "lib": ["dom", "dom.iterable", "esnext"],
+    "allowJs": true,
     "checkJs": true,
     "skipLibCheck": true,
     "strict": true,
@@ -18,5 +18,6 @@
     "incremental": true,
     "noUncheckedIndexedAccess": true
   },
+  "include": [".eslintrc.js", "prettier.config.cjs"],
   "exclude": ["node_modules"]
 }
